@@ -22,7 +22,7 @@ package tech.jhipster.config;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.web.cors.CorsConfiguration;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -1054,22 +1054,6 @@ public class JHipsterProperties {
 
             public void setPort(int port) {
                 this.port = port;
-            }
-
-            /*
-                queueSize is deprecated in favour of ringBufferSize
-             */
-            @Deprecated
-            public int getQueueSize() {
-                return getRingBufferSize();
-            }
-
-            /*
-                queueSize is deprecated in favour of ringBufferSize
-             */
-            @Deprecated
-            public void setQueueSize(int queueSize) {
-                setRingBufferSize(queueSize);
             }
 
             public int getRingBufferSize() {
