@@ -19,9 +19,9 @@
 
 package tech.jhipster.config.h2;
 
-import javax.servlet.Servlet;
-import javax.servlet.ServletContext;
-import javax.servlet.ServletRegistration;
+import jakarta.servlet.Servlet;
+import jakarta.servlet.ServletContext;
+import jakarta.servlet.ServletRegistration;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.sql.SQLException;
@@ -77,7 +77,7 @@ public class H2ConfigurationHelper {
     }
 
     /**
-     * Init the H2 console via H2's webserver when no servletContext {@link javax.servlet.ServletContext}
+     * Init the H2 console via H2's webserver when no servletContext {@link jakarta.servlet.ServletContext}
      * is available.
      */
     public static void initH2Console() {
@@ -85,7 +85,7 @@ public class H2ConfigurationHelper {
     }
 
     /**
-     * Init the H2 console via H2's webserver when no servletContext {@link javax.servlet.ServletContext}
+     * Init the H2 console via H2's webserver when no servletContext {@link jakarta.servlet.ServletContext}
      * is available.
      *
      * @param propertiesLocation the location where to find .h2.server.properties
@@ -109,14 +109,14 @@ public class H2ConfigurationHelper {
     /**
      * <p>initH2Console.</p>
      *
-     * @param servletContext a {@link javax.servlet.ServletContext} object.
+     * @param servletContext a {@link jakarta.servlet.ServletContext} object.
      */
     public static void initH2Console(ServletContext servletContext) {
         try {
             // We don't want to include H2 when we are packaging for the "prod" profile and won't
             // actually need it, so we have to load / invoke things at runtime through reflection.
             ClassLoader loader = Thread.currentThread().getContextClassLoader();
-            Class<?> servletClass = Class.forName("org.h2.server.web.WebServlet", true, loader);
+            Class<?> servletClass = Class.forName("org.h2.server.web.JakartaWebServlet", true, loader);
             Servlet servlet = (Servlet) servletClass.getDeclaredConstructor().newInstance();
 
             ServletRegistration.Dynamic h2ConsoleServlet = servletContext.addServlet("H2Console", servlet);
@@ -125,10 +125,10 @@ public class H2ConfigurationHelper {
             h2ConsoleServlet.setLoadOnStartup(1);
 
         } catch (ClassNotFoundException | LinkageError | NoSuchMethodException | InvocationTargetException e) {
-            throw new RuntimeException("Failed to load and initialize org.h2.server.web.WebServlet", e);
+            throw new RuntimeException("Failed to load and initialize org.h2.server.web.JakartaWebServlet", e);
 
         } catch (IllegalAccessException | InstantiationException e) {
-            throw new RuntimeException("Failed to instantiate org.h2.server.web.WebServlet", e);
+            throw new RuntimeException("Failed to instantiate org.h2.server.web.JakartaWebServlet", e);
         }
     }
 }
