@@ -21,8 +21,8 @@ package tech.jhipster.web.filter;
 
 import tech.jhipster.config.JHipsterProperties;
 
-import javax.servlet.*;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.*;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
 
