@@ -19,9 +19,9 @@
 
 package tech.jhipster.service.demo;
 
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.ManyToOne;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.ManyToOne;
 
 @Entity
 public class ChildEntity extends BaseEntity {
