@@ -19,9 +19,9 @@
 
 package tech.jhipster.service.demo;
 
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.OneToMany;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.OneToMany;
 import java.util.HashSet;
 import java.util.Set;
 
