@@ -20,8 +20,8 @@
 package tech.jhipster.config.apidoc;
 
 import io.swagger.v3.oas.models.OpenAPI;
-import org.springdoc.core.SpringDocConfiguration;
 
+import org.springdoc.core.configuration.SpringDocConfiguration;
 import tech.jhipster.config.JHipsterProperties;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
@@ -33,7 +33,7 @@ import org.springframework.context.annotation.Profile;
 import static tech.jhipster.config.JHipsterConstants.SPRING_PROFILE_API_DOCS;
 
 /**
- * OpenApi OpenAPI configuration.
+ * OpenAPI configuration.
  * <p>
  * Warning! When having a lot of REST endpoints, OpenApi can become a performance issue.
  * In that case, you can use the "no-api-docs" Spring profile, so that this bean is ignored.
