@@ -19,24 +19,19 @@
 
 package tech.jhipster.config.apidoc;
 
-import static org.springdoc.core.Constants.DEFAULT_GROUP_NAME;
-import static org.springdoc.core.Constants.SPRINGDOC_SHOW_ACTUATOR;
-import static org.springdoc.core.SpringDocUtils.getConfig;
-
 import io.swagger.v3.oas.models.info.Info;
 import java.nio.ByteBuffer;
 import java.util.List;
 import java.util.Optional;
+
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springdoc.core.GroupedOpenApi;
-import org.springdoc.core.GroupedOpenApi.Builder;
-import org.springdoc.core.SpringDocConfigProperties;
 import org.springdoc.core.customizers.ActuatorOpenApiCustomizer;
 import org.springdoc.core.customizers.ActuatorOperationCustomizer;
-import org.springdoc.core.customizers.OpenApiCustomiser;
+import org.springdoc.core.customizers.OpenApiCustomizer;
 import org.springdoc.core.customizers.OperationCustomizer;
+import org.springdoc.core.models.GroupedOpenApi;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
@@ -46,6 +41,10 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import tech.jhipster.config.JHipsterProperties;
 import tech.jhipster.config.apidoc.customizer.JHipsterOpenApiCustomizer;
+
+import static org.springdoc.core.utils.Constants.DEFAULT_GROUP_NAME;
+import static org.springdoc.core.utils.Constants.SPRINGDOC_SHOW_ACTUATOR;
+import static org.springdoc.core.utils.SpringDocUtils.getConfig;
 
 /**
  * OpenApi Groups configuration.
@@ -97,20 +96,18 @@ public class JHipsterSpringDocGroupsConfiguration {
     @Bean
     @ConditionalOnMissingBean(name = "openAPIDefaultGroupedOpenAPI")
     public GroupedOpenApi openAPIDefaultGroupedOpenAPI(
-        List<OpenApiCustomiser> openApiCustomisers,
+        List<OpenApiCustomizer> openApiCustomizers,
         List<OperationCustomizer> operationCustomizers,
-        @Qualifier("apiFirstGroupedOpenAPI") Optional<GroupedOpenApi> apiFirstGroupedOpenAPI,
-        SpringDocConfigProperties springDocConfigProperties
-    ) {
+        @Qualifier("apiFirstGroupedOpenAPI") Optional<GroupedOpenApi> apiFirstGroupedOpenAPI) {
         log.debug("Initializing JHipster OpenApi default group");
-        Builder builder = GroupedOpenApi.builder()
+        GroupedOpenApi.Builder builder = GroupedOpenApi.builder()
             .group(DEFAULT_GROUP_NAME)
             .pathsToMatch(properties.getDefaultIncludePattern());
-        openApiCustomisers.stream()
-            .filter(customiser -> !(customiser instanceof ActuatorOpenApiCustomizer))
-            .forEach(builder::addOpenApiCustomiser);
+        openApiCustomizers.stream()
+            .filter(customizer -> !(customizer instanceof ActuatorOpenApiCustomizer))
+            .forEach(builder::addOpenApiCustomizer);
         operationCustomizers.stream()
-            .filter(customiser -> !(customiser instanceof ActuatorOperationCustomizer))
+            .filter(customizer -> !(customizer instanceof ActuatorOperationCustomizer))
             .forEach(builder::addOperationCustomizer);
         apiFirstGroupedOpenAPI.map(GroupedOpenApi::getPackagesToScan)
             .ifPresent(packagesToScan -> packagesToScan.forEach(builder::packagesToExclude));
@@ -127,20 +124,16 @@ public class JHipsterSpringDocGroupsConfiguration {
     @ConditionalOnMissingBean(name = "openAPIManagementGroupedOpenAPI")
     @ConditionalOnProperty(SPRINGDOC_SHOW_ACTUATOR)
     public GroupedOpenApi openAPIManagementGroupedOpenAPI(
-        @Value("${spring.application.name:application}") String appName,
-        ActuatorOpenApiCustomizer actuatorOpenApiCustomiser,
-        ActuatorOperationCustomizer actuatorCustomizer
+        @Value("${spring.application.name:application}") String appName
     ) {
         log.debug("Initializing JHipster OpenApi management group");
         return GroupedOpenApi.builder()
             .group(MANAGEMENT_GROUP_NAME)
-            .addOpenApiCustomiser(openApi -> openApi.info(new Info()
+            .addOpenApiCustomizer(openApi -> openApi.info(new Info()
                 .title(StringUtils.capitalize(appName) + " " + MANAGEMENT_TITLE_SUFFIX)
                 .description(MANAGEMENT_DESCRIPTION)
                 .version(properties.getVersion())
             ))
-            .addOpenApiCustomiser(actuatorOpenApiCustomiser)
-            .addOperationCustomizer(actuatorCustomizer)
             .pathsToMatch(properties.getManagementIncludePattern())
             .build();
     }
